% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssanova_plots.R
\name{combineCuts}
\alias{combineCuts}
\title{combine different ssanova fits}
\usage{
combineCuts(..., cutnames, cutcolname = "timepoint")
}
\arguments{
\item{...}{data frames (from tongue.ssanova)}

\item{cutnames}{- the names to be used (e.g c("start", "middle", "end"))}

\item{cutcolname}{- the name of the column that will be used to specify the cut point.}
}
\value{
A combined data frame with an extra column
}
\description{
combine different ssanova fits
}
\seealso{
tongue.ssanova
}

