% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSSFF.R
\name{exportFiducialSSFF}
\alias{exportFiducialSSFF}
\title{Convert the text file containing spline data to ssff. This exports all the frames, 
which isn't really necessary}
\usage{
exportFiducialSSFF(txtfile = NULL, targdir = "./", missingAsZero = TRUE,
  tracksuf = "", filesuf = ".spln", fixPrompts = TRUE)
}
\arguments{
\item{txtfile}{the exported AAA spline file}

\item{targdir}{where the output will be written}

\item{missingAsZero}{Change NA to 0 so emu doesn't complain}

\item{tracksuf}{suffix appended to each filename to match the soundfiles already exported}
}
\description{
Convert the text file containing spline data to ssff. This exports all the frames, 
which isn't really necessary
}
\examples{
\dontrun{
exportFiducialSSFF("c:/AAA/Richard_BitePlane_Export_02_05_2015.txt", targdir="c:/Tabain/English_Ultrasound/")
}
}

