% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportSSFF.R
\name{exportSplinesSSFF}
\alias{exportSplinesSSFF}
\title{Convert the text file containing spline data to ssff}
\usage{
exportSplinesSSFF(txtfile = NULL, targdir = "./", missingAsZero = TRUE,
  tracksuf = "", fixPrompts = TRUE)
}
\arguments{
\item{txtfile}{the exported AAA spline file}

\item{targdir}{where the output will be written}

\item{missingAsZero}{Change NA to 0 so emu doesn't complain}

\item{tracksuf}{suffix appended to each filename to match the soundfiles already exported}

\item{fixPrompts}{attempts to use regular expressions to fix problems caused by tabs in the prompt list}
}
\value{
the list of prompts, invisibly
}
\description{
Convert the text file containing spline data to ssff
}
\examples{
\dontrun{
exportSplinesSSFF("c:/AAA/Richard_Spline_Export_02_05_2015.txt", targdir="c:/Tabain/English_Ultrasound/")
}
}

