% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_plots.R
\name{extractTongue}
\alias{extractTongue}
\title{Stuff to plot data in ssff files}
\usage{
extractTongue(trackdata)
}
\arguments{
\item{trackdata}{the track from emu.track}
}
\value{
A dataframe with X, Y, certainty, time and sample columns
}
\description{
Stuff to plot data in ssff files
}
\details{
Rearranges the trackdata into a sensible for for plotting
Not really needed for general use. Only works on a single token
}

