% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssanova_plots.R
\name{fit.ssanova}
\alias{fit.ssanova}
\title{Function for fitting the ssanova and returning a prediction
with confidence intervals.}
\usage{
fit.ssanova(DF, theformula, weight = NULL)
}
\arguments{
\item{DF}{dataframe with tongue spline}
}
\description{
Function for fitting the ssanova and returning a prediction
with confidence intervals.
}

