% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportSSFF.R
\name{fixNamesForEMU}
\alias{fixNamesForEMU}
\title{Create modified names for AAA files. Removes punctuation that offends tcltk}
\usage{
fixNamesForEMU(files)
}
\arguments{
\item{files}{the output of list.files}
}
\value{
list of modified names
}
\description{
Create modified names for AAA files. Removes punctuation that offends tcltk
}
\examples{
\dontrun{
f<-list.files(path="c:/Tabain/English_Ultrasound/", full.names=TRUE)
f1 <- gsub("_Track1", "", f)
f2 <- fixNamesForEMU(f1)
need.to.do <- basename(f) != basename(f2)
file.rename(from=f[need.to.do], to=f2[need.to.do])
}
}

