% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarstuff.R
\name{mkCartesianAll}
\alias{mkCartesianAll}
\title{Transform a typical ssanova back to
cartesian coordinates for plotting}
\usage{
mkCartesianAll(df)
}
\description{
Transform a typical ssanova back to
cartesian coordinates for plotting
}

