% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameAAA.R
\name{renameAAA}
\alias{renameAAA}
\title{Rename the files exported by AAA based on the contents of the .txt - lastname_firstname_word}
\usage{
renameAAA(txtfile = NULL)
}
\arguments{
\item{name}{of a file in the exported folder. If not provided then a file chooser will be used}
}
\value{
The original names and matching output
}
\description{
This function renames files exported by AAA. Input is the path to the folder
containing exported files, or select one of the files in the folder with a chooser dialog.
Renaming proceeds without warning. Likely to be fragile if there is lots of punctuation in the
word list.
}

