% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline_plots.R
\name{shinyTongue}
\alias{shinyTongue}
\title{Interactive visualization of splines by token}
\usage{
shinyTongue(splineDat)
}
\arguments{
\item{splineDat}{output from splineDatForGG}
}
\value{
Nothing - a shiny app should appear in a browser
}
\description{
Interactive visualization of splines by token
}
\examples{
\dontrun{
labs <- paste(labsminus1.marija,labs.marija,labsplus1.marija)
g<-splineDatForGG(tongue.marija, labs)
shinyTongue(g)
}
}

