% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline_plots.R
\name{splineDatForGG}
\alias{splineDatForGG}
\title{Extract splines for all tokens in a trackdata file.}
\usage{
splineDatForGG(trackdata, tags)
}
\arguments{
\item{trackdata}{the trackdata from emu.track}

\item{tags}{the labels that you want to use. Probably paste(labsminus1, labs, labsplus1)}
}
\value{
a big dataframe for visualization
}
\description{
Extract splines for all tokens in a trackdata file.
}
\examples{
\dontrun{
labs <- paste(labsminus1.marija,labs.marija,labsplus1.marija)
g<-splineDatForGG(tongue.marija, labs)

ggplot(g, aes(x=X,y=Y,group=interaction(factor(sample),unique.tag), colour=time)) + geom_line() +
geom_point(aes(size=certainty)) + facet_wrap(~unique.tag)
ggplot(subset(g, orig.tag=="h a: d"), aes(x=X,y=Y,group=interaction(factor(sample),unique.tag), colour=time)) + geom_line() +
geom_point(aes(size=certainty)) + facet_wrap(~unique.tag)
}
}

