% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssanova_plots.R
\name{tongue.ssanova}
\alias{tongue.ssanova}
\title{Fit ssanova to groups of tongue splines}
\usage{
tongue.ssanova(tonguedat, labels, individual = FALSE)
}
\arguments{
\item{tonguedat}{- the tongue data, usually with emu.track and a cut
option}

\item{labels}{- the labels used to define grouped splines}

\item{individual}{- return all labels (including repeats), if true 
possibly good for debugging.}
}
\value{
a data frame with smoothed values and CI suitable for
plotting with ggplot2
}
\description{
Fit ssanova to groups of tongue splines
}
\examples{
\dontrun{
plotoptions <- list(
 scale_x_continuous(breaks=seq(from=-50,to=50,by=20)),
  scale_y_continuous(breaks=seq(from=-40, to=30,by=20))
  )
  
gg <- tongue.ssanova(tongue.data.cut_0.5, labs.segs)
ggplot(gg, aes(x=X, y=Y, group=orig.tag, colour=orig.tag)) + 
geom_path() + geom_smooth(aes(ymin=Y-CI, ymax=Y+CI), stat='identity') +
 coord_fixed() + theme_bw()
 
g0 <- tongue.ssanova(tongue.data.cut_0.0, labs.segs)
g1 <- tongue.ssanova(tongue.data.cut_0.5, labs.segs)
g2 <- tongue.ssanova(tongue.data.cut_1.0, labs.segs)

gall <- combineCuts(g0, g1, g2, cutnames=c("start", "middle", "end"))
ggplot(gall, aes(x=X, y=Y, group=interaction(orig.tag,timepoint), colour=orig.tag)) + 
geom_path() + geom_smooth(aes(ymin=Y-CI, ymax=Y+CI), stat='identity') +
 coord_fixed() + theme_bw() + plotoptions
 
 
ggplot(gall, aes(x=X, y=Y, group=interaction(orig.tag,timepoint), colour=timepoint)) + 
geom_path() + geom_smooth(aes(ymin=Y-CI, ymax=Y+CI), stat='identity') +
 facet_wrap(~orig.tag) + coord_fixed() + theme_bw() + plotoptions 
 
}
}

