% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULTRA.R
\name{ULTRA}
\alias{ULTRA}
\title{ULTRA}
\usage{
ULTRA(data, namebin)
}
\arguments{
\item{data}{dataframe containing z-scaled continuous features and single binary outcome}

\item{namebin}{name of the binary outcome variable}
}
\value{
Displays a combined cowplot. Note the individual plots can be accessed from the object returned by the ULTRA function.
}
\description{
Fits prediction model for all features and selected features.
}
