% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULTRA.R
\name{ULTRAplot}
\alias{ULTRAplot}
\title{ULTRAplot}
\usage{
ULTRAplot(results)
}
\arguments{
\item{results}{The object returned by the ULTRA function.}
}
\value{
Visual output displaying a combined cowplot. Note the individual plots can be accessed from the object returned by the ULTRA function.
}
\description{
Plots the AUC from the prediction model for all features and selected features.
}
